+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [UNRELEASED] - YYYY-MM-DD
+### ✨ Added
+- Add counts dialog for events and annotations to show the number of unique event/annotation types ([#483](https://github.com/cbrnr/mnelab/pull/483) by [Clemens Brunner](https://github.com/cbrnr))
+
 ### 🌀 Changed
 - Streamline settings (by using INI format on all platforms) as well as the settings dialog ([#482](https://github.com/cbrnr/mnelab/pull/482) by [Clemens Brunner](https://github.com/cbrnr))